pyproject.toml
[project]
name = "nlp-qwen-ocr"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
	"torch>=2.2",
	"transformers>=4.44.0",
	"qwen-vl-utils==0.0.8",
	"pillow>=10.4.0",
	"python-docx>=1.1.0",
	"tqdm>=4.66.0",
]
